Recalculate weekly intetest fee, add calculateAprIncrement helper
