+++ src/pages/RefinancePage/components/RefinanceTable/TableCells/APRCell.tsx
@@ -1,11 +1,10 @@
 import { FC } from 'react'
 
-import moment from 'moment'
-
 import { Loan } from '@banx/api/core'
 import { HealthColorDecreasing, convertAprToApy, getColorByPercent } from '@banx/utils'
 
 import { INCREASE_PERCENT_APR_PER_HOUR, MAX_APY_INCREASE_PERCENT } from '../constants'
+import { calculateAprIncrement } from '../helpers'
 
 import styles from '../RefinanceTable.module.less'
 
@@ -14,18 +13,11 @@ interface APRCellProps {
 }
 
 export const APRCell: FC<APRCellProps> = ({ loan }) => {
-  const { amountOfBonds } = loan.bondTradeTransaction
-  const { refinanceAuctionStartedAt } = loan.fraktBond
+  const aprIncrement = calculateAprIncrement(loan)
 
-  const currentTime = moment()
-  const auctionStartTime = moment.unix(refinanceAuctionStartedAt)
-  const hoursSinceStart = currentTime.diff(auctionStartTime, 'hours')
+  const colorAPR = getColorByPercent(aprIncrement, HealthColorDecreasing)
 
-  const apr = amountOfBonds / 100 + hoursSinceStart
-
-  const colorAPR = getColorByPercent(apr, HealthColorDecreasing)
-
-  const apy = Math.min(convertAprToApy(apr / 100), MAX_APY_INCREASE_PERCENT)
+  const apy = Math.min(convertAprToApy(aprIncrement / 100), MAX_APY_INCREASE_PERCENT)
 
   const isApyIncreaseRateVisible = apy < MAX_APY_INCREASE_PERCENT
